rename api/data-mgt/ folder to api/data/
