+++ src/app/app.service.spec.ts
@@ -5,7 +5,7 @@
 
 import { TestBed } from '@angular/core/testing';
 import { LocalMemorySessionApiModule } from 'src/app/api/session/local-memory/local-memory-session-api.module';
-import { LocalMemoryDataModule } from 'src/app/api/data-mgt/local-memory/local-memory-data-api.module';
+import { LocalMemoryDataModule } from 'src/app/api/data/local-memory/local-memory-data-api.module';
 import { AppService } from './app.service';
 
 describe('AppService', () => {
