+++ src/app/management/management.service.spec.ts
@@ -6,7 +6,7 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { LocalMemorySessionApiModule } from 'src/app/api/session/local-memory/local-memory-session-api.module';
-import { LocalMemoryDataModule } from 'src/app/api/data-mgt/local-memory/local-memory-data-api.module';
+import { LocalMemoryDataModule } from 'src/app/api/data/local-memory/local-memory-data-api.module';
 import { ManagementService } from './management.service';
 
 describe('ManagementService', () => {
