+++ src/app/management/routes/data-manager.service-directive.ts
@@ -6,7 +6,7 @@
 import { Directive, OnDestroy } from '@angular/core';
 import { BehaviorSubject, from, Observable, of, Subject, ReplaySubject } from 'rxjs';
 import { catchError, delay, finalize, map, mapTo, mergeMap, startWith, tap, toArray } from 'rxjs/operators';
-import { EntityDataApiIService } from 'src/app/api/data-mgt/entity-data-api.iservice';
+import { EntityDataApiIService } from 'src/app/api/data/entity-data-api.iservice';
 import { AbstractEntity } from 'src/app/models/AbstractEntity';
 import { AuthorizedAccess } from 'src/app/models/AuthorizedAccess';
