+++ src/app/management/routes/sellers/form-dialog/seller-manager-form.service.spec.ts
@@ -4,7 +4,7 @@
 // https://opensource.org/licenses/MIT
 
 import { TestBed } from '@angular/core/testing';
-import { LocalMemoryDataModule } from 'src/app/api/data-mgt/local-memory/local-memory-data-api.module';
+import { LocalMemoryDataModule } from 'src/app/api/data/local-memory/local-memory-data-api.module';
 import { SellerManagerFormService } from './seller-manager-form.service';
 
 describe('SellerManagerFormService', () => {
